Refactor registration class into RegisterDoctor and RegisterPatient classes
